+++ src/components/Project/Resource/DFR3MappingPage.tsx
@@ -20,6 +20,7 @@ import { ProjectSidebar } from "@app/components/Project/ProjectSidebar";
 import { useAppDispatch } from "@app/store/hooks";
 
 import DFR3Icon from "@mui/icons-material/ShowChart";
+import Snackbar from "@mui/joy/Snackbar";
 import { AddFromServiceDialog } from "@app/components/Project/Resource/AddFromServiceDialog";
 
 const DFR3MappingPage = (): JSX.Element => {
@@ -73,6 +74,24 @@ const DFR3MappingPage = (): JSX.Element => {
         appDispatch(addDFR3MappingToProject({ projectId, dfr3Mappings: [resource] }));
         setOpenAddDFR3MappingFromServiceDialog(false);
     };
+
+    // snackbar
+    const [snackbarOpen, setSnackbarOpen] = React.useState<boolean>(false);
+    const [snackbarMessage, setSnackbarMessage] = React.useState<string>("");
+    const [snackbarColor, setSnackbarColor] = React.useState<"success" | "danger" | "warning" | "neutral">("neutral");
+    const success = useSelector((state: RootState) => state.project.success);
+    const error = useSelector((state: RootState) => state.project.error);
+    React.useEffect(() => {
+        if (error) {
+            setSnackbarMessage(`Error: ${error}`);
+            setSnackbarColor("danger");
+            setSnackbarOpen(true);
+        } else if (success) {
+            setSnackbarMessage(success);
+            setSnackbarColor("success");
+            setSnackbarOpen(true);
+        }
+    }, [success, error]);
 
     return (
         <Container sx={{ display: "flex", flexDirection: "column", height: "100vh" }} maxWidth="xl">
@@ -132,6 +151,19 @@ const DFR3MappingPage = (): JSX.Element => {
                     </>
                 )}
             </Box>
+            <Snackbar
+                anchorOrigin={{ vertical: "bottom", horizontal: "left" }}
+                open={snackbarOpen}
+                onClose={() => {
+                    setSnackbarOpen(false);
+                    setSnackbarMessage("");
+                }}
+                variant="outlined"
+                color={snackbarColor}
+                autoHideDuration={2000}
+            >
+                {snackbarMessage}
+            </Snackbar>
         </Container>
     );
 };