+++ src/reducer/projectSlice.ts
@@ -20,7 +20,8 @@ const initialState: ProjectState = {
     projectVisualizations: <Visualization[]>[],
     deletedVisualizationIds: <string[]>[],
     loading: false,
-    error: null
+    error: null,
+    success: null
 };
 
 export const getProjects = createAsyncThunk(
@@ -398,6 +399,7 @@ const projectSlice = createSlice({
             .addCase(getProjects.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(getProjects.fulfilled, (state, action) => {
                 state.loading = false;
@@ -411,6 +413,7 @@ const projectSlice = createSlice({
             .addCase(searchProjects.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(searchProjects.fulfilled, (state, action) => {
                 state.loading = false;
@@ -424,10 +427,12 @@ const projectSlice = createSlice({
             .addCase(createProject.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(createProject.fulfilled, (state, action) => {
                 state.loading = false;
                 state.project = action.payload;
+                state.success = "Successfully created the project";
             })
             .addCase(createProject.rejected, (state, action) => {
                 state.loading = false;
@@ -437,6 +442,7 @@ const projectSlice = createSlice({
             .addCase(getProject.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(getProject.fulfilled, (state, action) => {
                 state.loading = false;
@@ -450,6 +456,7 @@ const projectSlice = createSlice({
             .addCase(getProjectDatasets.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(getProjectDatasets.fulfilled, (state, action) => {
                 state.loading = false;
@@ -463,6 +470,7 @@ const projectSlice = createSlice({
             .addCase(searchProjectDatasets.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(searchProjectDatasets.fulfilled, (state, action) => {
                 state.loading = false;
@@ -476,10 +484,12 @@ const projectSlice = createSlice({
             .addCase(addDatasetToProject.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(addDatasetToProject.fulfilled, (state, action) => {
                 state.loading = false;
                 state.projectDatasets = action.payload?.datasets;
+                state.success = "Successfully added datasets to the project";
             })
             .addCase(addDatasetToProject.rejected, (state, action) => {
                 state.loading = false;
@@ -489,10 +499,12 @@ const projectSlice = createSlice({
             .addCase(deleteProjectDatasets.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(deleteProjectDatasets.fulfilled, (state, action) => {
                 state.loading = false;
                 state.deletedDatasetIds = action.payload;
+                state.success = "Successfully deleted the project datasets";
             })
             .addCase(deleteProjectDatasets.rejected, (state, action) => {
                 state.loading = false;
@@ -502,6 +514,7 @@ const projectSlice = createSlice({
             .addCase(getProjectWorkflows.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(getProjectWorkflows.fulfilled, (state, action) => {
                 state.loading = false;
@@ -514,10 +527,12 @@ const projectSlice = createSlice({
             .addCase(deleteProjectWorkflows.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(deleteProjectWorkflows.fulfilled, (state, action) => {
                 state.loading = false;
                 state.deletedWorkflowIds = action.payload;
+                state.success = "Successfully deleted the project workflows";
             })
             .addCase(deleteProjectWorkflows.rejected, (state, action) => {
                 state.loading = false;
@@ -527,6 +542,7 @@ const projectSlice = createSlice({
             .addCase(searchProjectWorkflows.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(searchProjectWorkflows.fulfilled, (state, action) => {
                 state.loading = false;
@@ -540,6 +556,7 @@ const projectSlice = createSlice({
             .addCase(getProjectHazards.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(getProjectHazards.fulfilled, (state, action) => {
                 state.loading = false;
@@ -553,10 +570,12 @@ const projectSlice = createSlice({
             .addCase(addHazardToProject.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(addHazardToProject.fulfilled, (state, action) => {
                 state.loading = false;
                 state.projectHazards = action.payload?.hazards;
+                state.success = "Successfully added hazards to the project";
             })
             .addCase(addHazardToProject.rejected, (state, action) => {
                 state.loading = false;
@@ -566,10 +585,12 @@ const projectSlice = createSlice({
             .addCase(deleteProjectHazards.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(deleteProjectHazards.fulfilled, (state, action) => {
                 state.loading = false;
                 state.deletedHazardIds = action.payload;
+                state.success = "Successfully deleted the project hazards";
             })
             .addCase(deleteProjectHazards.rejected, (state, action) => {
                 state.loading = false;
@@ -579,6 +600,7 @@ const projectSlice = createSlice({
             .addCase(searchProjectHazards.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(searchProjectHazards.fulfilled, (state, action) => {
                 state.loading = false;
@@ -592,6 +614,7 @@ const projectSlice = createSlice({
             .addCase(getProjectDRF3Mappings.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(getProjectDRF3Mappings.fulfilled, (state, action) => {
                 state.loading = false;
@@ -605,10 +628,12 @@ const projectSlice = createSlice({
             .addCase(addDFR3MappingToProject.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(addDFR3MappingToProject.fulfilled, (state, action) => {
                 state.loading = false;
                 state.projectDFR3Mappings = action.payload?.dfr3Mappings;
+                state.success = "Successfully added dfr3mappings to the project";
             })
             .addCase(addDFR3MappingToProject.rejected, (state, action) => {
                 state.loading = false;
@@ -618,10 +643,12 @@ const projectSlice = createSlice({
             .addCase(deleteProjectDFR3Mappings.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(deleteProjectDFR3Mappings.fulfilled, (state, action) => {
                 state.loading = false;
                 state.deletedDFR3MappingIds = action.payload;
+                state.success = "Successfully deleted the project DFR3 Mappings";
             })
             .addCase(deleteProjectDFR3Mappings.rejected, (state, action) => {
                 state.loading = false;
@@ -631,6 +658,7 @@ const projectSlice = createSlice({
             .addCase(searchProjectDFR3Mappings.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(searchProjectDFR3Mappings.fulfilled, (state, action) => {
                 state.loading = false;
@@ -644,6 +672,7 @@ const projectSlice = createSlice({
             .addCase(getProjectVisualizations.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(getProjectVisualizations.fulfilled, (state, action) => {
                 state.loading = false;
@@ -656,10 +685,12 @@ const projectSlice = createSlice({
             .addCase(deleteProjectVisualizations.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(deleteProjectVisualizations.fulfilled, (state, action) => {
                 state.loading = false;
                 state.deletedVisualizationIds = action.payload;
+                state.success = "Successfully deleted the project visualizations";
             })
             .addCase(deleteProjectVisualizations.rejected, (state, action) => {
                 state.loading = false;
@@ -668,10 +699,12 @@ const projectSlice = createSlice({
             .addCase(createProjectVisualization.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(createProjectVisualization.fulfilled, (state, action) => {
                 state.loading = false;
                 state.projectVisualizations = action.payload?.visualizations;
+                state.success = "Successfully created the project visualizations";
             })
             .addCase(createProjectVisualization.rejected, (state, action) => {
                 state.loading = false;
@@ -680,10 +713,12 @@ const projectSlice = createSlice({
             .addCase(addLayerToVisualization.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(addLayerToVisualization.fulfilled, (state, action) => {
                 state.loading = false;
                 state.projectVisualizations = action.payload?.visualizations;
+                state.success = "Successfully added layers to the project visualizations";
             })
             .addCase(addLayerToVisualization.rejected, (state, action) => {
                 state.loading = false;
@@ -693,6 +728,7 @@ const projectSlice = createSlice({
             .addCase(searchProjectVisualizations.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(searchProjectVisualizations.fulfilled, (state, action) => {
                 state.loading = false;
@@ -706,10 +742,12 @@ const projectSlice = createSlice({
             .addCase(deleteProject.pending, (state) => {
                 state.loading = true;
                 state.error = null;
+                state.success = null;
             })
             .addCase(deleteProject.fulfilled, (state, action) => {
                 state.loading = false;
                 state.deletedProjectId = action.payload;
+                state.success = "Successfully deleted the project";
             })
             .addCase(deleteProject.rejected, (state, action) => {
                 state.loading = false;